ADD PWM ok

+++ MM32/mLLD/TIM/PWM/MM32_PWM.h
@@ -2,7 +2,7 @@
  * @Author: IlleniumDillon 147900130@qq.com
  * @Date: 2022-12-13 18:30:02
  * @LastEditors: IlleniumDillon 147900130@qq.com
- * @LastEditTime: 2022-12-13 19:20:11
+ * @LastEditTime: 2022-12-14 10:34:17
  * @FilePath: \CODE\MM32\mLLD\TIM\PWM\MM32_PWM.H
  * @Description: 这是默认设置,请设置`customMade`, 打开koroFileHeader查看配置 进行设置: https://github.com/OBKoro1/koro1FileHeader/wiki/%E9%85%8D%E7%BD%AE
  */
@@ -52,8 +52,28 @@ typedef struct
     MM32PWM_Channel channel;
 }MM32PWM_Pin;
 
-void MM32PWM_pinInit(MM32PWM_Pin* pin, uint32_t freq);
+/**
+ * @description: 初始化PWM引脚
+ * @param {MM32PWM_Pin*} pin PWM引脚结构体指针
+ * @param {float} freq PWM频率
+ * @return {*}
+ * @example
+ * \code
+ *    MM32PWM_pinInit(&pwmpin,1000);
+ * \endcode
+ */
+void MM32PWM_pinInit(MM32PWM_Pin* pin, float freq);
 
+/**
+ * @description: 设定PWM占空比
+ * @param {MM32PWM_Pin*} pin PWM引脚结构体指针
+ * @param {float} duty 占空比
+ * @return {*}
+ * @example
+ * \code
+ *    MM32PWM_setDuty(&pwmpin,0.5);
+ * \endcode
+ */
 void MM32PWM_setDuty(MM32PWM_Pin* pin, float duty);
 
 #endif
